requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ollama==0.1.7
pydantic==1.10.13  # ✅ NÃO precisa de Rust
python-multipart==0.0.6
python-dotenv==1.0.0
# psycopg2-binary==2.9.7 --only-binary=all  # ✅ Força binário pré-compilado
